pyproject.toml
[project]
name = "yijian-zhi-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.5.5",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-openai>=0.3.19",
    "langchain-qdrant>=0.2.0",
    "python-docx>=1.1.2",
    "unstructured>=0.17.2",
]

[dependency-groups]
dev = [
    "pyright>=1.1.401",
    "pytest>=8.3.5",
    "ruff>=0.11.8",
]

[tool.pyright]
typeCheckingMode = "standard"
include = [
    ".",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "ASYNC", "UP"]
